+++ src/contract/CMakeLists.txt
@@ -23,12 +23,12 @@ set(CONTRACT_HEADERS
   ${CMAKE_SOURCE_DIR}/src/contract/variables/safeunorderedmap.h
   ${CMAKE_SOURCE_DIR}/src/contract/variables/safebool.h
   ${CMAKE_SOURCE_DIR}/src/contract/erc20.h
-  ${CMAKE_SOURCE_DIR}/src/contract/dexv2pair.h
-  ${CMAKE_SOURCE_DIR}/src/contract/dexv2factory.h
-  ${CMAKE_SOURCE_DIR}/src/contract/dexv2library.h
+  ${CMAKE_SOURCE_DIR}/src/contract/dexv2/dexv2pair.h
+  ${CMAKE_SOURCE_DIR}/src/contract/dexv2/dexv2factory.h
+  ${CMAKE_SOURCE_DIR}/src/contract/dexv2/dexv2library.h
   ${CMAKE_SOURCE_DIR}/src/contract/customcontracts.h
-  ${CMAKE_SOURCE_DIR}/src/contract/dexv2router02.h
-  ${CMAKE_SOURCE_DIR}/src/contract/uq112x112.h
+  ${CMAKE_SOURCE_DIR}/src/contract/dexv2/dexv2router02.h
+  ${CMAKE_SOURCE_DIR}/src/contract/dexv2/uq112x112.h
   PARENT_SCOPE
 )
 
@@ -39,10 +39,10 @@ set(CONTRACT_SOURCES
   ${CMAKE_SOURCE_DIR}/src/contract/dynamiccontract.cpp
   ${CMAKE_SOURCE_DIR}/src/contract/nativewrapper.cpp
   ${CMAKE_SOURCE_DIR}/src/contract/erc20.cpp
-  ${CMAKE_SOURCE_DIR}/src/contract/dexv2pair.cpp
-  ${CMAKE_SOURCE_DIR}/src/contract/dexv2factory.cpp
-  ${CMAKE_SOURCE_DIR}/src/contract/dexv2library.cpp
-  ${CMAKE_SOURCE_DIR}/src/contract/dexv2router02.cpp
+  ${CMAKE_SOURCE_DIR}/src/contract/dexv2/dexv2pair.cpp
+  ${CMAKE_SOURCE_DIR}/src/contract/dexv2/dexv2factory.cpp
+  ${CMAKE_SOURCE_DIR}/src/contract/dexv2/dexv2library.cpp
+  ${CMAKE_SOURCE_DIR}/src/contract/dexv2/dexv2router02.cpp
   ${CMAKE_SOURCE_DIR}/src/contract/erc20wrapper.cpp
   PARENT_SCOPE
 )
