+++ tests/contract/dexv2.cpp
@@ -1,8 +1,8 @@
 #include "../../src/libs/catch2/catch_amalgamated.hpp"
 #include "../../src/contract/contractmanager.h"
-#include "../../src/contract/dexv2pair.h"
-#include "../../src/contract/dexv2factory.h"
-#include "../../src/contract/dexv2router02.h"
+#include "../../src/contract/dexv2/dexv2pair.h"
+#include "../../src/contract/dexv2/dexv2factory.h"
+#include "../../src/contract/dexv2/dexv2router02.h"
 #include "../../src/contract/abi.h"
 #include "../../src/utils/db.h"
 #include "../../src/utils/options.h"
